# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/contest/admin/add": {
            "post": {
                "description": "添加比赛管理",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contest"
                ],
                "summary": "添加比赛管理",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/contract.AddContestAdminRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.AddContestAdminResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/contract.AddContestAdminResponse"
                        }
                    }
                }
            }
        },
        "/contest/admin/delete": {
            "delete": {
                "description": "删除比赛管理",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contest"
                ],
                "summary": "删除比赛管理",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/contract.DeleteContestAdminRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.DeleteContestAdminResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/contract.DeleteContestAdminResponse"
                        }
                    }
                }
            }
        },
        "/contest/create": {
            "post": {
                "description": "创建比赛",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contest"
                ],
                "summary": "创建比赛",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/contract.CreateContestRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.CreateContestResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/contract.CreateContestResponse"
                        }
                    }
                }
            }
        },
        "/contest/generate": {
            "post": {
                "description": "生成比赛用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contest"
                ],
                "summary": "生成比赛用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/contract.GenerateParticipantsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.GenerateParticipantsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/contract.GenerateParticipantsResponse"
                        }
                    }
                }
            }
        },
        "/contest/join": {
            "post": {
                "description": "加入比赛",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contest"
                ],
                "summary": "加入比赛",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/contract.JoinContestRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.JoinContestResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/contract.JoinContestResponse"
                        }
                    }
                }
            }
        },
        "/contest/list": {
            "get": {
                "description": "比赛列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contest"
                ],
                "summary": "比赛列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/contract.ListContestsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.ListContestsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/contract.ListContestsResponse"
                        }
                    }
                }
            }
        },
        "/contest/problem/add": {
            "post": {
                "description": "添加比赛题目",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contest"
                ],
                "summary": "添加比赛题目",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/contract.AddContestProblemRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.AddContestProblemResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/contract.AddContestProblemResponse"
                        }
                    }
                }
            }
        },
        "/contest/problem/delete": {
            "delete": {
                "description": "删除比赛题目",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contest"
                ],
                "summary": "删除比赛题目",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/contract.DeleteContestProblemRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.DeleteContestProblemResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/contract.DeleteContestProblemResponse"
                        }
                    }
                }
            }
        },
        "/contest/rank": {
            "post": {
                "description": "比赛排名",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contest"
                ],
                "summary": "比赛排名",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/contract.ContestRankRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.ContestRankResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/contract.ContestRankResponse"
                        }
                    }
                }
            }
        },
        "/contest/search": {
            "post": {
                "description": "比赛搜索",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contest"
                ],
                "summary": "比赛搜索",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/contract.SearchContestRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.SearchContestResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/contract.SearchContestResponse"
                        }
                    }
                }
            }
        },
        "/contest/show": {
            "get": {
                "description": "比赛详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contest"
                ],
                "summary": "比赛详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/contract.ShowContestRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.ShowContestResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/contract.ShowContestResponse"
                        }
                    }
                }
            }
        },
        "/contest/update": {
            "post": {
                "description": "比赛更新",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contest"
                ],
                "summary": "比赛更新",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/contract.UpdateContestRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.UpdateContestResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/contract.UpdateContestResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "contract.AddContestAdminRequest": {
            "type": "object",
            "required": [
                "contest_id",
                "user_id"
            ],
            "properties": {
                "contest_id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "contract.AddContestAdminResponse": {
            "type": "object",
            "properties": {
                "base_response": {
                    "$ref": "#/definitions/contract.BaseResponse"
                }
            }
        },
...
        "contract.CreateContestRequest": {
            "type": "object",
            "required": [
                "end_time",
                "start_time",
                "title"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "end_time": {
                    "type": "integer"
                },
                "problem_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "start_time": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "contract.CreateContestResponse": {
            "type": "object",
            "properties": {
                "base_response": {
                    "$ref": "#/definitions/contract.BaseResponse"
                },
                "contest": {
                    "$ref": "#/definitions/entity.Contest"
                }
            }
        },
        "contract.DeleteContestAdminRequest": {
            "type": "object",
            "required": [
                "contest_id",
                "user_id"
            ],
            "properties": {
                "contest_id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "contract.DeleteContestAdminResponse": {
            "type": "object",
            "properties": {
                "base_response": {
                    "$ref": "#/definitions/contract.BaseResponse"
                }
            }
        },
...
        "contract.GenerateParticipantsRequest": {
            "type": "object",
            "required": [
                "contest_id",
                "generate_count"
            ],
            "properties": {
                "contest_id": {
                    "type": "integer"
                },
                "generate_count": {
                    "type": "integer"
                }
            }
        },
...
    }
}